+++ packages/carbon-web-components/src/components/slug/slug.ts
@@ -145,8 +145,13 @@ export default class CDSSlug extends CDSToggleTip {
     name === 'revert-active' ? this.parentElement?.requestUpdate() : ``;
   }
 
-  updated(changedProperties) {
-    super.updated(changedProperties);
+  // connectedCallback() {
+  //   super.connectedCallback();
+  //   this.setAttribute('autoAlign', '');
+  // }
+
+  updated() {
+    super.updated();
     if (
       this.kind !== SLUG_KIND.HOLLOW &&
       this.dotType !== SLUG_DOT_TYPE.HOLLOW