+++ packages/web-components/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.3.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-web-components@2.3.0-rc.1...@carbon/ibmdotcom-web-components@2.3.0) (2024-02-12)
+
+
+### Bug Fixes
+
+* **footer:** locale modal button now working ([#11529](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/issues/11529)) ([c062fbb](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/commit/c062fbbea5f30113c1b38648242044a53a695440))
+
+
+
+
+
+# [2.3.0-rc.1](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-web-components@2.3.0-rc.0...@carbon/ibmdotcom-web-components@2.3.0-rc.1) (2024-02-09)
+
+**Note:** Version bump only for package @carbon/ibmdotcom-web-components
+
+
+
+
+
 # [2.3.0-rc.0](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/compare/@carbon/ibmdotcom-web-components@2.2.0-rc.0...@carbon/ibmdotcom-web-components@2.3.0-rc.0) (2024-02-02)